+++ src/main/java/com/phantomwing/rusticdelight/util/ComposterHelper.java
@@ -26,7 +26,7 @@ public class ComposterHelper {
         );
 
         // 65% chance
-        registerCompostableItems(0.5f,
+        registerCompostableItems(0.65f,
                 ModItems.BELL_PEPPER_GREEN,
                 ModItems.BELL_PEPPER_YELLOW,
                 ModItems.BELL_PEPPER_RED,
@@ -34,5 +34,16 @@ public class ComposterHelper {
                 ModItems.WILD_COTTON,
                 ModItems.WILD_BELL_PEPPERS
         );
+
+        // 85% chance
+        registerCompostableItems(0.85f,
+                ModItems.CHERRY_BLOSSOM_COOKIE,
+                ModItems.CHERRY_BLOSSOM_CHEESECAKE_SLICE
+        );
+
+        // 100% chance
+        registerCompostableItems(1.0f,
+                ModItems.CHERRY_BLOSSOM_CHEESECAKE
+                );
     }
 }